Update 'AmountFromValue' & 'dAmount' value
Update to ensure that a numeric is being passed to 'AmountFromValue' to return a more descriptive type error when the input parameter is not numeric. While this isn't a functional change (since non-numeric input values would previously fail anyways, returning only an "Invalid amount" error message) this clarifies the response message and allows for a faster exit response by not needing to do subsequent mathematical or external function checks.

Update 'dAmount' value parameter to be in scope with RSTR supply given that the previous value parameter was too small and would throw "Invalid amount" error if an amount was passed greater than the previous value set.